requirements.txt
# aiohappyeyeballs==2.6.1
# aiohttp==3.13.2
# aiosignal==1.4.0
# annotated-types==0.7.0
# asgiref==3.10.0
# attrs==25.4.0
# bitarray==3.8.0
# certifi==2025.11.12
# charset-normalizer==3.4.4
# ckzg==2.1.5
# cytoolz==1.1.0
# Django==5.2.7
# django-cors-headers==4.9.0
# djangorestframework==3.16.1
# eth-account==0.13.7
# eth-hash==0.7.1
# eth-keyfile==0.8.1
# eth-keys==0.7.0
# eth-rlp==2.2.0
# eth-typing==5.2.1
# eth-utils==5.3.1
# eth_abi==5.2.0
# filelock==3.19.1
# frozenlist==1.8.0
# fsspec==2025.9.0
# hexbytes==1.3.1
# idna==3.11
# Jinja2==3.1.6
# joblib==1.5.2
# MarkupSafe==2.1.5
# mpmath==1.3.0
# multidict==6.7.0
# networkx==3.5
# numpy==2.3.5
# pandas==2.3.3
# parsimonious==0.10.0
# propcache==0.4.1
# pycryptodome==3.23.0
# pydantic==2.12.5
# pydantic_core==2.41.5
# python-dateutil==2.9.0.post0
# python-decouple==3.8
# pytz==2025.2
# pyunormalize==17.0.0
# pywin32==311
# regex==2025.11.3
# requests==2.32.5
# rlp==4.1.0
# scikit-learn==1.7.2
# scipy==1.16.3
# setuptools==70.2.0
# six==1.17.0
# sqlparse==0.5.3
# sympy==1.14.0
# threadpoolctl==3.6.0
# toolz==1.1.0
# torch==2.9.1+cpu
# types-requests==2.32.4.20250913
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# tzdata==2025.2
# urllib3==2.5.0
# web3==7.14.0
# websockets==15.0.1
# xgboost==3.1.2
# yarl==1.22.0

asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-decouple==3.8
sqlparse==0.5.3

# ML for honeypot_ML
xgboost==1.7.6
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2

# ML for exit_ML (attention MIL model)
torch==2.9.1+cpu
python-dateutil==2.9.0.post0
--extra-index-url https://download.pytorch.org/whl/cpu

# Data collection
web3==6.20.0
requests==2.32.3
